requirements.txt
Django
djangorestframework
mysqlclient
django-cors-headers
djangorestframework-simplejwt
Pillow
